split LinearMemory into a new file
